+++ pkg/daemon/libp2p.go
@@ -1,0 +1,82 @@
+// Copyright 2021 Trim21<trim21.me@gmail.com>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+package daemon
+
+import (
+	"context"
+
+	"github.com/ipfs/go-datastore"
+	"github.com/ipfs/go-ipns"
+	"github.com/libp2p/go-libp2p"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	"github.com/libp2p/go-libp2p-core/host"
+	"github.com/libp2p/go-libp2p-core/pnet"
+	"github.com/libp2p/go-libp2p-core/routing"
+	dht "github.com/libp2p/go-libp2p-kad-dht"
+	"github.com/libp2p/go-libp2p-kad-dht/dual"
+	record "github.com/libp2p/go-libp2p-record"
+	"github.com/multiformats/go-multiaddr"
+	"github.com/pkg/errors"
+)
+
+func SetupLibp2p(
+	ctx context.Context,
+	hostKey crypto.PrivKey,
+	secret pnet.PSK,
+	listenAddrs []multiaddr.Multiaddr,
+	ds datastore.Batching,
+	opts ...libp2p.Option,
+) (host.Host, *dual.DHT, error) {
+	var ddht *dual.DHT
+	var err error
+	finalOpts := []libp2p.Option{
+		libp2p.Identity(hostKey),
+		libp2p.ListenAddrs(listenAddrs...),
+		libp2p.PrivateNetwork(secret),
+		libp2p.Routing(func(h host.Host) (routing.PeerRouting, error) {
+			ddht, err = newDHT(ctx, h, ds)
+
+			return ddht, err
+		}),
+	}
+	finalOpts = append(finalOpts, opts...)
+
+	h, err := libp2p.New(
+		ctx,
+		finalOpts...,
+	)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "failed init libp2p")
+	}
+
+	return h, ddht, nil
+}
+
+func newDHT(ctx context.Context, h host.Host, ds datastore.Batching) (*dual.DHT, error) {
+	dhtOpts := []dual.Option{
+		dual.DHTOption(dht.NamespacedValidator("pk", record.PublicKeyValidator{})),
+		dual.DHTOption(dht.NamespacedValidator("ipns", ipns.Validator{KeyBook: h.Peerstore()})),
+		dual.DHTOption(dht.Concurrency(dhtConcurrency)),
+		dual.DHTOption(dht.Mode(dht.ModeAuto)),
+	}
+	if ds != nil {
+		dhtOpts = append(dhtOpts, dual.DHTOption(dht.Datastore(ds)))
+	}
+
+	d, err := dual.New(ctx, h, dhtOpts...)
+
+	return d, errors.Wrap(err, "failed to init dualdht")
+}
